NEW FORMAT FOR LOG FILES
The filament_simulator node now writes the log files in a new format.
What does this mean for the user? It means the simulations are much faster to run, and most of the time will take up significantly less space on disk (although this depends on how many wind snapshots you are using).
Files with the old format can no longer be generated, as it would be a hassle to maintain the code for both.
However, the Gaden_player node IS backwards compatible, so any simulations you have run in the past can still be played back.

The test_env folder has received some changes as well, since the format for the parameters that determine where to store and read log files (results_location and simulation_data) are *slightly* different now (see the included launch files for working examples).