Added automatic  modification of .world files
Automatically fixes the size and center pose of the floor map according to the coordinates and dimensions of the environment.  To use it, set the path to the world file as the parameter "worldFile" in the prerocessing.launch

Kind of fragile, because it does not really parse the contents, but instead overwrites the entire floorMap section. Still, I am pretty sure it's going to do more good than bad., and you can always turn it off by leaving the parameter empty.