2.0.0
Gaden is now officially working in ROS2! In fact, from this point onwards the ROS2 version (Gaden 2) is the recommended version of Gaden.

Bug fixes will still be released for Gaden 1 for a while, but it will receive no new features. At some point in the future, it might stop receiving support at all.